Revert back to an older version.
Learning the importance of using git every step of the way.  Had to use
time machine to find the right version of my code that worked.  Tried to
do this with git and had some issues.  Saved my code that didn't work
with the nw_312 in this filename.  Going back and building on the add
item I recovered from time machine.