Do Over…reverted back to some older code.
After adding a bunch of complexity to my code that was basically
functioning, I decided to go back to my latest working version.  Setup
a variable in the validate form function that stores true or false as
an array for each form element.  In the end, if all fields are not
valid, the function will list the ones that need correction. This and
prevent the function from storing the data.